use shared_ptr to forward declaration of MultigridState
