+++ include/ginkgo/core/solver/solver_base.hpp
@@ -71,13 +71,11 @@ enum class initial_guess_mode {
 
 
 namespace multigrid {
-namespace experimental {
 
 
 class MultigridState;
 
 
-}  // namespace experimental
 }  // namespace multigrid
 
 
@@ -88,7 +86,7 @@ class MultigridState;
  */
 class ApplyWithInitialGuess {
 protected:
-    friend class multigrid::experimental::MultigridState;
+    friend class multigrid::MultigridState;
 
     /**
      * Applies a linear operator to a vector (or a sequence of vectors) with
@@ -167,7 +165,7 @@ private:
 template <typename DerivedType>
 class EnableApplyWithInitialGuess : public ApplyWithInitialGuess {
 protected:
-    friend class multigrid::experimental::MultigridState;
+    friend class multigrid::MultigridState;
 
     explicit EnableApplyWithInitialGuess(
         initial_guess_mode guess = initial_guess_mode::provided)
